+++ app/src/main/java/com/example/weatherjourney/core/model/weather/CurrentWeather.kt
@@ -2,9 +2,9 @@ package com.example.weatherjourney.core.model.weather
 
 data class CurrentWeather(
     val date: String,
-    val temp: Double,
-    val windSpeed: Double,
-    val humidity: Double,
-    val pressure: Double,
+    val temp: Float,
+    val windSpeed: Float,
+    val humidity: Float,
+    val pressure: Float,
     val weatherType: WeatherType,
 )
