+++ app/src/main/java/com/example/weatherjourney/core/model/weather/HourlyWeather.kt
@@ -2,7 +2,7 @@ package com.example.weatherjourney.core.model.weather
 
 data class HourlyWeather(
     val date: String,
-    val temp: Double,
-    val windSpeed: Double,
+    val temp: Float,
+    val windSpeed: Float,
     val weatherType: WeatherType,
 )
